[#42] - replace the query trait with an injected query service and users repository
